requirements.txt
opencv-python
openai
pydub
numpy
TTS
python-dotenv
aiohttp
aiofiles
runware
gradio